router and bindings definition
